requirements.txt
Faker==9.3.1
mysql_connector_repackaged==0.3.1
selenium==4.0.0
